+++ mlos_bench/mlos_bench/environments/local/local_env.py
@@ -18,7 +18,6 @@ from types import TracebackType
 from typing import Any, Dict, Iterable, List, Mapping, Optional, Tuple, Type, Union
 from typing_extensions import Literal
 
-import pytz
 import pandas
 
 from mlos_bench.environments.status import Status
@@ -28,7 +27,7 @@ from mlos_bench.services.base_service import Service
 from mlos_bench.services.types.local_exec_type import SupportsLocalExec
 from mlos_bench.tunables.tunable import TunableValue
 from mlos_bench.tunables.tunable_groups import TunableGroups
-from mlos_bench.util import path_join
+from mlos_bench.util import datetime_parser, path_join
 
 _LOG = logging.getLogger(__name__)
 
@@ -216,44 +215,6 @@ class LocalEnv(ScriptEnv):
             data.rename(str.rstrip, axis='columns', inplace=True)
         return data
 
-    # All timestamps in the telemetry data must be greater than this date
-    # (a very rough approximation for the start of this feature).
-    _MIN_TS = datetime(2024, 1, 1, 0, 0, 0, tzinfo=pytz.UTC)
-
-    @staticmethod
-    def _datetime_parser(datetime_col: pandas.Series) -> pandas.Series:
-        """
-        Attempt to convert a column to a datetime format.
-
-        Parameters
-        ----------
-        datetime_col : pandas.Series
-            The column to convert.
-
-        Returns
-        -------
-        pandas.Series
-            The converted datetime column.
-
-        Raises
-        ------
-        ValueError
-            On parse errors.
-        """
-        new_datetime_col = pandas.to_datetime(datetime_col, utc=False)
-        # If timezone data is missing, assume the local timezone.
-        if new_datetime_col.dt.tz is None:
-            local_tzinfo = datetime.now().astimezone().tzinfo
-            new_datetime_col = new_datetime_col.dt.tz_localize(local_tzinfo)
-        assert new_datetime_col.dt.tz is not None
-        # And convert it to UTC.
-        new_datetime_col = new_datetime_col.dt.tz_convert('UTC')
-        if new_datetime_col.isna().any():
-            raise ValueError(f"Invalid date format in the telemetry data: {datetime_col}")
-        if new_datetime_col.le(LocalEnv._MIN_TS).any():
-            raise ValueError(f"Invalid date range in the telemetry data: {datetime_col}")
-        return new_datetime_col
-
     def status(self) -> Tuple[Status, datetime, List[Tuple[datetime, str, Any]]]:
 
         (status, timestamp, _) = super().status()
@@ -271,7 +232,7 @@ class LocalEnv(ScriptEnv):
 
             data = self._normalize_columns(
                 pandas.read_csv(fname, index_col=False))
-            data.iloc[:, 0] = self._datetime_parser(data.iloc[:, 0])
+            data.iloc[:, 0] = datetime_parser(data.iloc[:, 0])
 
             expected_col_names = ["timestamp", "metric", "value"]
             if len(data.columns) != len(expected_col_names):
@@ -281,7 +242,7 @@ class LocalEnv(ScriptEnv):
                 # Assume no header - this is ok for telemetry data.
                 data = pandas.read_csv(
                     fname, index_col=False, names=expected_col_names)
-                data.iloc[:, 0] = self._datetime_parser(data.iloc[:, 0])
+                data.iloc[:, 0] = datetime_parser(data.iloc[:, 0])
 
         except FileNotFoundError as ex:
             _LOG.warning("Telemetry CSV file not found: %s :: %s", self._read_telemetry_file, ex)
